requirements.txt
Django==5.0.1
djangorestframework==3.14.0
pillow==10.2.0
django-oauth-toolkit==2.3.0
python-decouple==3.8